+++ Math/Parallelogram.cs
@@ -8,12 +8,12 @@ namespace BlokMath
 {
     internal class Parallelogram
     {
-        public float a { get; private set; }
-        public float b { get; private set; }
-        public float a_h { get; private set; }
-        public float a_b { get; private set; }
+        public decimal a { get; private set; }
+        public decimal b { get; private set; }
+        public decimal a_h { get; private set; }
+        public decimal a_b { get; private set; }
 
-        public float d1 { get; private set; }
-        public float d2 { get; private set; }
+        public decimal d1 { get; private set; }
+        public decimal d2 { get; private set; }
     }
 }
