+++ Math/Trapezoid.cs
@@ -8,15 +8,15 @@ namespace BlokMath
 {
     internal class Trapezoid
     {
-        public float a { get; private set; }
-        public float b { get; private set; }
+        public decimal a { get; private set; }
+        public decimal b { get; private set; }
 
-        public float c { get; private set; }
-        public float d { get; private set; }
+        public decimal c { get; private set; }
+        public decimal d { get; private set; }
 
-        public float h { get; private set; }
+        public decimal h { get; private set; }
 
-        public float d1 { get; private set; }
-        public float d2 { get; private set; }
+        public decimal d1 { get; private set; }
+        public decimal d2 { get; private set; }
     }
 }
